+++ tests/Unit/MeiliSearchEngineTest.php
@@ -123,11 +123,46 @@ class MeiliSearchEngineTest extends TestCase
         $client = m::mock(Client::class);
         $engine = new MeiliSearchEngine($client);
 
-        $results = $engine->mapIds([
+        $results = $engine->mapIdsFrom([
             'nbHits' => 0, 'hits' => [],
-        ]);
+        ], 'id');
 
         $this->assertEquals(0, count($results));
+    }
+
+    public function test_map_ids_returns_correct_values_of_primary_key()
+    {
+        $client = m::mock(Client::class);
+        $engine = new MeiliSearchEngine($client);
+
+        $results = $engine->mapIdsFrom([
+            'nbHits' => 5,
+            'hits' => [
+                [
+                    'some_field' => 'something',
+                    'id' => 1,
+                ],
+                [
+                    'some_field' => 'foo',
+                    'id' => 2,
+                ],
+                [
+                    'some_field' => 'bar',
+                    'id' => 3,
+                ],
+                [
+                    'some_field' => 'baz',
+                    'id' => 4,
+                ],
+            ],
+        ], 'id');
+
+        $this->assertEquals($results->all(), [
+            1,
+            2,
+            3,
+            4,
+        ]);
     }
 
     public function test_map_correctly_maps_results_to_models()
@@ -240,10 +275,13 @@ class MeiliSearchEngineTest extends TestCase
     {
         $client = m::mock(Client::class);
         $client->shouldReceive('index')->with('table')->andReturn($index = m::mock(Indexes::class));
-        $index->shouldReceive('addDocuments')->with([['id' => 'my-meilisearch-key.1']], 'id');
+        $index->shouldReceive('addDocuments')->once()->with([[
+            'meilisearch-key' => 'my-meilisearch-key.5',
+            'id' => 5,
+        ]], 'meilisearch-key');
 
         $engine = new MeiliSearchEngine($client);
-        $engine->update(Collection::make([new MeiliSearchCustomKeySearchableModel()]));
+        $engine->update(Collection::make([new MeiliSearchCustomKeySearchableModel(['id' => 5])]));
     }
 
     public function test_flush_a_model_with_a_custom_meilisearch_key()
@@ -390,5 +428,10 @@ class MeiliSearchCustomKeySearchableModel extends SearchableModel
     public function getScoutKey()
     {
         return 'my-meilisearch-key.'.$this->getKey();
+    }
+
+    public function getScoutKeyName()
+    {
+        return 'meilisearch-key';
     }
 }